+++ app/src/main/java/com/playground/navigationwithtabs/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.playground.navigationwithtabs
+package com.playground.navigationwithtabs.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.playground.navigationwithtabs.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
