+++ app/src/main/res/layout/fragment_view_pager.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ViewPagerFragment">
+    tools:context=".ui.ViewPagerFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
